split rerun-64bit-tests from test_base
